modificado:     lib/CiA402Device 	modificado:     lib/fcontrol 	modificado:     lib/sensor-integration 	modificado:     main/clInc20degstep.cpp 	modificado:     main/frontierstest.cpp 	modificado:     main/frontierstest_id.cpp 	nuevo archivo:  main/iros2020-identification.cpp 	modificado:     main/ol-pos-360.cpp 	nuevo archivo:  tools/test_sensor.cpp

+++ main/clInc20degstep.cpp
@@ -23,7 +23,7 @@ int main ()
 {
     //--sensor--
     SerialArduino tilt;
-    float incSensor,oriSensor;
+    double incSensor,oriSensor;
 //    sleep(4); //wait for sensor
 
     ofstream data("/home/humasoft/code/graficas/graficas_demos/clinc20degs-400g.csv",std::ofstream::out); // /home/humasoft/code/graficas
